chore: moved format error and format response
