+++ src/client/base-client.ts
@@ -64,37 +64,4 @@ export abstract class BaseClient<T extends {}> {
   public isInitialized() {
     return this.initialized;
   }
-
-  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-  protected formatResponse<G extends Record<string | number, any>, O>(
-    response: FetchResponse<G, O, "application/json">,
-  ): DataResponseValue<G, O> {
-    let outputData = undefined;
-    let outputMetadata = undefined;
-    if (!response) {
-      return {} as DataResponseValue<G, O>;
-    }
-    if (response.data) {
-      const dataEntries = Object.entries(response.data);
-      // by docs the body should contain a "data"
-      // properties with the response values
-      const containsData = dataEntries.find((x) => x[0] === "data");
-      const containsMetadata = dataEntries.find((x) => x[0] === "metadata");
-
-      if (containsMetadata) {
-        outputMetadata = containsMetadata[1];
-      }
-      outputData = containsData ? containsData[1] : response.data;
-    }
-
-    return {
-      data: outputData,
-      metadata: outputMetadata,
-      error: response.error,
-    } as unknown as DataResponseValue<G, O>;
-  }
-
-  protected formatError<G, O>(reason: unknown): DataResponseValue<G, O> {
-    return { error: reason } as unknown as DataResponseValue<G, O>;
-  }
 }
